Category layout update, Timers from database in WatchOverview, Joda Time implemented.
Updated the category layout with:
	* number of watches;
	* currently running watches;
	* new placeholder image for logo.

The Watches/Timers are now obtained from the database and putted in the view when entering a category. The timers don't run yet and it isn't possible to pause timers, or make new timers without the application throwing a NPE.

Implemented Joda Time. A library which gives me the ability to subtract Dates from each other to calculate the current time.